InduCred: android app to set photon's credentials
This app reads the SSID of your phone, asks for the network password and translates the info to an audible msg, mosre or modem style. The message is played on the phones loudspeaker.
If the Photon has the inducred.ino loaded and running it will it can pickup the sound through a coil of a 5 volt relay connected to A4.